Use kubectl from path instead of relying on legacy verify-kube-binaries.
