+++ kubetest2-gce/deployer/down.go
@@ -69,8 +69,6 @@ func (d *deployer) verifyDownFlags() error {
 		return err
 	}
 
-	d.kubectlPath = filepath.Join(d.RepoRoot, "cluster", "kubectl.sh")
-
 	if d.GCPProject == "" {
 		return fmt.Errorf("gcp project must be set")
 	}
