+++ kubetest2-gce/deployer/up.go
@@ -25,6 +25,37 @@ import (
 	"sigs.k8s.io/kubetest2/pkg/exec"
 )
 
+func (d *deployer) IsUp() (up bool, err error) {
+	klog.V(1).Info("GCE deployer starting IsUp()")
+
+	if err := d.init(); err != nil {
+		return false, fmt.Errorf("isUp failed to init: %s", err)
+	}
+
+	if d.GCPProject == "" {
+		return false, fmt.Errorf("isup requires a GCP project")
+	}
+
+	env := d.buildEnv()
+	// naive assumption: nodes reported = cluster up
+	// similar to other deployers' implementations
+	args := []string{
+		d.kubectlPath,
+		"get",
+		"nodes",
+		"-o=name",
+	}
+	cmd := exec.Command(args[0], args[1:]...)
+	cmd.SetEnv(env...)
+	cmd.SetStderr(os.Stderr)
+	lines, err := exec.OutputLines(cmd)
+	if err != nil {
+		return false, fmt.Errorf("is up failed to get nodes: %s", err)
+	}
+
+	return len(lines) > 0, nil
+}
+
 func (d *deployer) Up() error {
 	klog.V(1).Info("GCE deployer starting Up()")
 
@@ -106,8 +137,6 @@ func (d *deployer) verifyUpFlags() error {
 	if err := d.setRepoPathIfNotSet(); err != nil {
 		return err
 	}
-
-	d.kubectlPath = filepath.Join(d.RepoRoot, "cluster", "kubectl.sh")
 
 	// verifyUpFlags does not check for a gcp project because it is
 	// assumed that one will be acquired from boskos if it is not set
